ignore .bundle/config and vendor/
There are various ways bundler can be configured for running and testing
the code in this repo, and each way depends on the context and
environment.  For example "bundle config path vendor/bundle" is one way
to stop bundle installing gems system-wide via sudo without the user's
permission, which is a bug in bundler described here:

  https://github.com/bundler/bundler/issues/3006

Another potentially useful setting is BUNDLE_DISABLE_SHARED_GEMS for
when you don't want to use system gems (i.e. ones installed via
packages).

The repo itself shouldn't be too opinionated about these settings, and
we don't want to track anything installed to vendor/ so it makes sense
to add these to .gitignore.