fix windows install by using IPv4 tftp symlinks
In the Hyper-V case we use name-remapping for the bootmgr.exe
using the \x expanding to the hexadecimal version of the IP-address.
With IPv6 now being used (even when connecting from a IPv4 addr),
this address and thus the remapped name is different than in Cloud5
and thus did not find our symlinks.