Sync config/boot.rb version specifications with Gemfile requirements
Assuming that rubygems follow semantic versioning guidelines,
we want to lock on patch level releases. It seems useful to me
to be consistent with the versioning in the Gemfile and this
part of crowbar to ensure that production and dev environments
always have exactly the same rubygem dependency restrictions.

Rebase of barclamp-crowbar/pull/1324