Fix apache syntax error for crowbar vhost (bsc#948717)
Crowbar needs now mod_proxy (and some other related modules) but
using these modules via the chef cookbook only triggers a a2enmod
which writes the module to /etc/sysconfig/apache2.
A reload is not enough because the listed modules from the sysconfig
file are added as parameters to the command line when starting apache.

https://bugzilla.suse.com/show_bug.cgi?id=948717