resolver: Reload dnsmasq after DNS change if node is dns-server
Since dnsmasq is caching its data, we need to reload the local dnsmasq
in case there's a change in the DNS config (when we run bind). This
makes it possible to have a node get a new IP address and have us still
be able to resolve to the correct IP address.

We use a delayed notification because bind is also using a delayed one,
so it makes no sense to reload dnsmasq before bind.