webui: Only show non-allocated admin IP address on discovery
When we do hardware-installing/hardware-installed, then the node already
has an allocated IP address, even if it's not necessarily using it yet:

 - when doing hardware-installing/hardware-installed outside of the
   discovery process (that can happen with the bios/raid barclamps),
   then yes, it just works fine. But arguable, not really a case that
   we're seeing for all practical purposes ;-)

 - in discovery, we have two things:
   - the network role is not used, so it won't set the IP address from
     the host range. So at first, it would seem that you are correct,
     but...
   - the dhcp configuration is actually updated, and on lease renewal,
     the node will get the allocated IP address.

All in all, we are in a gray area here, and it pretty much depends on
the timing. That being said, I would argue that people generally won't
want to connect to the discovered node at this point in time (they would
connect before allocating it) and what we display in the UI here is
consistent with what is put in the DNS config.