bind9: Have discovered nodes in DNS
Since discovered nodes don't have an allocated IP address in the admin
range, but just an IP address from the DHCP pool, we need to deal with
them in a special way.

Also, we need to keep in mind that the IP address from the DHCP pool can
be used for several nodes over time; especially if a node is discovered
and then turned off for a long period of time, another discovered node
could very well re-use the same IP address. So in that case, we take the
node that was discovered the most recently for that IP address.