webui: Rework repo controller and how we save repo in data bag
We now only use one data bag, with a structure similar to the one from
the yaml file.

Also the controller is slightly more intelligent, being able to deal
with repos that don't exist, re-saving a repo that had changes on
enabling, not returning an error when enabling a repo that is already
eanbled, etc.