webui: Go back to one data bag item per repository
It is well documented in the chef documentation that data bags cannot be
concurrently edited. By having all repos in one data bag item, we could
hit a situation with concurrent requests enabling/disabling two
different repos, and the two requests would conflict, with possibly one
losing the change from the other because it loaded the repo too early
(before the other saved) and saves with stale data.

By moving to one data bag item per repository, the only conflict we
could have is when two concurrent requests edit the same repo, which is
something that cannot be resolved anyway. So it's safer.