Introduce new attribute to request openvswitch installation (bsc#950798)
Some nodes, e.g. compute nodes when using ovs+gre/vxlan, need openvswitch
running without requiring any bridges been created by the barclamp. This
commit decouples the installion of the packages and startup of the service
from the "add_ovs_bridge" attribute. And adds the new "needs_openvswitch"
attribute. This attribute in not present in the proposal as it's specific
to single nodes and not for the whole deployment of a network proposal.

https://bugzilla.suse.com/show_bug.cgi?id=950798