Renew DHCP lease after hardware-installing transtion
During the "hardware-installing" transition a new admin IP Address (from the
"host" ranges) gets allocated to the node. Force a DHCP renewal to change to
the new IP Address directly after the transition. This avoids having the IP
Address changed at a later (unpredictable) time during the remainder of the
sledgehammer phase.