Fix ifcfg template for openvswitch device
According to the SLES12-SP1 documentation STARTMODE can be "auto" for
OvsBridges as well.
We need the OVS_BRIDGE='yes' and OVS_BRIDGE_PORT_DEVICE_* settings as wicked
will remove the bridge during shutdown and recreate them at boottime. (Which
went unnoticed until recently because of a bug in the openvswitch service
dependencies)