Added method to disable all non-core Proposals
This is needed to keep the non-core Cookbooks from being executed when a
node is coming back to "ready" after performing and OS upgrade. It's supposed
to be called before running the actual upgrade (i.e. before calling
prepare_nodes_for_os_upgrade()). Unfortunately we can't use
ServiceObject.destroy_active() for this, because that would cause (queued)
chef-client runs on the non-upgrades nodes. We don't want that here.