Drop piechart from dashboard
Adapting the piechart to the new view/ajax way is to much work the
current card. For now lets drop it due to it gives anyway not much
important information. A user with less notes can see it while looking
at the nodes and a user with ~100 nodes can't see the difference in the
piechart.

When a user asks for it we should reimplement this in a better way.