crowbar: Fix node runlist sorting to be predictable
If two roles have the same priority, there was nothing that was
guaranteeing a fixed order. So sort by priority and then name.

This bug could result in issues for HA clusters where one member of the
cluster runs one role while others are running another role with the
same priority -- triggering timeouts on the sync marks.