Enable parameter wrapping for json
This change is needed that we can send json requests with params without
a namespace. For html requests it is recommended to namespace them but
for json requests it is in general not wanted because it makes the api
requests more complex. The rails wrapper just ensures that both ways
works for json.