Leave upgrade related role for nodes in crowbar_upgrade state.
Currently when the proposal is applied for nodes with crowbar_upgrade
state, we have to make sure (in apply_role_pre_chef_call) that such nodes
won't lose the state (otherwise it would be changed back to ready
at the end of proposal commit).

But if node would not have any upgrade related role, it
would not reach this code in apply_role_pre_chef_call. As we
allow also non-upgrade related roles in the runlist of nodes in
crowbar_upgrade state, commiting proposal with only these (non-upgrade)
roles would mean that the state of node would be restored to ready.