crowbar: Switch from puma to passenger (bsc#963738)
We're hitting issues with the worker & thread model of puma:

 - with a config for 5 workers and 1 thread per worker, for some reason,
   we don't have concurrent requests going through, resulting in a poor
   user experience (just one tab working) and breakage for some bits
   (deploying xen compute roles, which require concurrency)

 - with a config with multiple threads per worker, we hit some
   SQLite3::BusyException way too often with no good reason (apparently,
   two threads blocking each other when acquiring the lock, and
   resulting in a deadlock -- which is really weird as one should win
   and not block...)

So let's move to passenger, which is widely used, and integrates
directly into apache (nice bonus).

Note that we don't use the passenger module for apache and instead keep
using a reverse proxy. This is because our cookbooks sometimes reload
(or even restart) apache, and in both cases, this kills passenger and
therefore the rails app. Which is quite bad if the result of the chef
run was important for the rails app...

https://bugzilla.suse.com/show_bug.cgi?id=963738