crowbar: Use temporary directory when exporting supportconfig
While not strictly needed, it's slightly better in order to avoid
non-crowbar owned files in the rails app tmp directory. We use a chown
at the end, of course, but if the rails app is stopped in the middle,
then the tmp directory contains files that can't be removed.

Using a proper temporary directory helps (since it may be cleaned up on
reboot, or by scripts).