crowbar: Fix deployment queue to not use outdated information (bsc#965610)
We were caching the content of the queue, which could result in some
outdated information being used.

For instance, when applying a proposal with an empty queue and applying
a second proposal a few seconds later on (which would go in the queue):
at the end of the apply_role of the first proposal, we still had the
cached empty queue and so were no processing the queue further instead
of reloading the queue and seeing a new proposal was queued.

https://bugzilla.suse.com/show_bug.cgi?id=965610