provisioner: Fix chef crash on admin server when forgetting nodes
I've seen that several times, but each time I didn't have time to debug
and this change was "good enough". It's still somehow correct because if
you read the code path that sets the interface list, it can be nil when
a conduit cannot be resolved, for instance.

Of course, now that I'm trying to reproduce, I'm not able to...