Reject incoming DNS requests while restoring/upgrading (bsc#966419)
This is needed to avoid a running cloud accessing the (incomplete) DNS info
that the admin node would handout when the installer runs during a restore (or
upgrade) as that might cause services to fail when upgrading the admin node.

Ideally we would reconfigure bind to only listen on 127.0.0.1 during that
stage, but that requires some more changes as dnsmasq is already running on
that (to be done in a later commit).

https://bugzilla.suse.com/show_bug.cgi?id=966419