provisioner: Fix crash on admin server when a node is missing ohai data
node[:macaddress] is the same as node.macaddress, but doesn't fail when
that attribute is missing. And that happens when (for some reason), ohai
data for that node is missing.

It's better to not crash the chef runs on the admin server (as this can
block the whole allocation process for other nodes), so just do this.