bind9: Do not create invalid zone when discovered node has no ohai data
Due to some bug, it can happen that a discovered node has some missing
ohai data. This is wrong of course and should be fixed.

But a result of this is that we create some invalid DNS zone, with
something that has a line like:

  d52-54-77-77-77-02 IN A

(and no address).

We don't want to break the DNS config, so protect against this.