Moved lock library into proper base class
In order to hopefully fix bsc#966590 I have moved the shared locking
base methods into a Lock::Base class, this is already done that way
within many ruby projects like rails and so on. This should hopefully
fix "Circular dependency detected while autoloading constant
Crowbar::Lock" issues. I can not be sure as this error is not
reproducable for me.

https://bugzilla.suse.com/show_bug.cgi?id=966590