call chef-client directly
single_chef_client.sh runs in the background, and though comes back
immediately. It is better to wait here and not go in the background.
The benefit here is that we can be sure that all necessary changes
after restore are applied, and the user can continue to use the cloud
immediately after restore.
This also helps unbreaking our CI where testsetup needs the chef-server
really quickly after restore is done.