during restore don't create a backup that already exists (bsc#966489)
it can happen that the user triggered a restore, the backup with name x
is not indexed in the db anymore, but it still exists in the filesystem.
so until restore is done, creating a backup with the same name should
be blocked.

https://bugzilla.suse.com/show_bug.cgi?id=966489
(cherry picked from commit 0a631a887a1f15d8c4acd3c17e2135ea727fd2e1)