Improve how we sort interfaces when there's no interface map
When there's no interface map, we sort interfaces by their name. Which
is kind of okay-ish, except that on first boot, the names are assigned
in a first-come first-serve basis. Which might not match the
expectation.

So instead, when there's no interface map, sort by bus order and
fallback on the name.