show sanity checks instead of failing at start-time
the rails app just refused to start when the sanity checks failed.
now the UI redirects to a sanity-check view and lists the errors.

for the future the checks can be extended to cover other areas than
just the network and thus make sure crowbar is in a sane state during
start-time.