Revert "don't cache memoized variable"
This reverts commit 84d01b7102e878810e2abd3260825be1767c41f1.

we actually rely on the same pathname object here, as we know it has
been extracted and upgraded there. if the tmpdir is somehow is gone
after the upgrade has been performed on the files, and extracted again,
then we lose the upgrade modifications.
the main point here is that we have to work with one tmpdir all the time

(cherry picked from commit d6a119ee872961bb3329c614b435b3977a7dcdaa)