provisioner: Fix reboot of discovered node
When a node is discovered, but not allocated, it is not assigned an IP
address in the admin network yet. Due to the dhcp template interpreting
"no IP" as "deny booting", this resulted in reboot of discovered nodes
blocking the nodes.

Now in such cases, we simply take the current IP address of the node.