provisioner: Fix reboot of discovered node (bsc#988021)
When a node is discovered, but not allocated, it is not assigned an IP
address in the admin network yet. Due to the dhcp template interpreting
"no IP" as "deny booting", this resulted in reboot of discovered nodes
blocking the nodes.

Now in such cases, we simply take the current IP address of the node.

https://bugzilla.suse.com/show_bug.cgi?id=988021

(cherry picked from commit 7711b4c63e1428bf8faeee5179da62878ebded1a)