make role recipes skip recipes during nova-ha-compute batch (bsc#1004758)
When we apply a proposal, we potentially have multiple batches of roles
to apply.  The batches are applied sequentially, but for each batch,
chef-client is run in parallel on all nodes affected by roles in that
batch.  (Typically there is only one role per batch, but barclamps such
as ipmi, logging, and ntp have batches with multiple roles.)

However for each chef-client run we still currently apply all
recipes in the run list.  This is incredibly wasteful, and has a huge
negative impact on the time taken to apply proposals, especially the
later ones, since by that time the run list is very long.

So introduce a mechanism which takes advantage of the new(ish) role
recipes in order to skip recipes which don't need to be run.  Currently,
the only time recipes are skipped is during application of the
nova-ha-compute batch, which is the 3rd batch in nova proposals.  This
should eliminate sync_mark drift between the chef-client runs during
this 3rd batch.  (There is no drift in the 1st batch, because we
reset sync marks before applying any proposal.)

  https://bugzilla.suse.com/show_bug.cgi?id=1004758

There is one exception: we never skip deployer-client, because it's
required both for heartbeat and setting up some utilities for recipes.