chef: Move inclusion of recipes from deployer-client to provisioner-base
The repos cookbook only matters outside of the discovery image, so fits
provisioner-base better.

The kernel-panic cookbook should be done by provisioner-base, but is
also useful in the discovery image, so reflect this.

Rationale for moving this is that deployer-client will always run in all
cases, even when we only run a subset of roles, so we should minimize
the amount of work it does in general.