[3.0] crowbar: Ensure run_list is how we want it to be
When we add a role to a node, we need the role in the run_list_map and
in the run_list, but if it was in one or the other, we were not doing
anything. This resulted in semi-broken setups (which could happen due to
crashes, or manual editing) not getting fixed when re-applying a
barclamp.

Now, when we change the run_list, we really ensure that it's how we want
it to be.

(cherry picked from commit b341f0b9b7ec70e6b2131f94ef3f39e6da60a89b)