crowbar: Fix queue message when applying a proposal (bsc#989958)
There was an attempt to fix the internal API so that
ServiceObject.apply_role always returns [int, str], but this actually
breaks the UI-only path in the controller.

So revert this and document with a FIXME that this should eventually be
fixed by moving everything to [int, hash] where we will have more
freedom.

https://bugzilla.suse.com/show_bug.cgi?id=989958
(cherry picked from commit 1ad066488776c313b1a01b3ddc19b216111e0658)