output useful debugging on unmatched interface (bsc#990745)
If we fail to find an interface matching a pattern such as "+10g1"
for a given node, give some useful debugging.  We don't make it a
warning, because there are cases where we expect this to happen,
e.g. with rarely used networks such as 'os_sdn'.

This helps more quickly identify incorrectly configured network.json
files, and was experienced in a scenario where the Crowbar nodes were a
mix of physical machines and VMs with slow virtio interfaces.

https://bugzilla.suse.com/show_bug.cgi?id=990745