fail with useful error on unmatched admin interface (bsc#990745)
Catch the case where we failed to find an admin interface for a node,
and bail with a more helpful error message.

This was experienced in a scenario where the Crowbar nodes were a mix of
physical machines and VMs with slow virtio interfaces, but the
network.json was incorrectly configured and the admin network conduit
failed to match the slow virtio interfaces.

https://bugzilla.suse.com/show_bug.cgi?id=990745