Force reconnect to the database
In case the database connection gets cut, ActiveRecord will not recognize
that the connection is broken until the next access attempt. In this case
it will raise the StatementInvalid error.
In our case this happens when we apply a proposal e.g. provisioner on
the admin node.

For more information please check:
https://github.com/cantino/huginn/issues/1116
https://github.com/rails/rails/issues/12867

The full error message is:
ActiveRecord::StatementInvalid: PG::ConnectionBad: PQconsumeInput()
server closed the connection unexpectedly