crowbar: Use network definitions from the network role
In the "hardware-installing" state nodes do not carry have the network
attributes yet (only after chef-client ran for the transition). Access the
network definition via the network proposal role in that case.

The check for the "bastion" network needs to be done on the role in any case.
Since "node.networks" will only return networks which are enabled already for
the node.