crowbar: Add conduit_resolver library
This library provides a central place for all the functions needed to
resolve network conduit names ("intf0", ...) and crowbar's internal
interface names ("1g1", "10g2", ...) to physical interface names of of a
node.

It's supposed to be usable by the crowbar rails app and from within chef
recipes. For this to work a small set for functions need to be
overwriten to provide an abstraction layer where needed:

- cr_network_config
  Returns the network config (from network barclamp)
- cr_ohai_network
  Returns the OHAI network attributes from the node
- cr_node_roles
  Returns the roles from the node
- cr_dmi_system
  Returns the DMI system attributes from the node
- cr_node_bond_list
  Return the list of bonds from the node
- cr_error(s)
  Output an error message