ohai: common ohai attribute for libvirt guest uuid
This adds a new attribute node[:crowbar_ohai][:libvirt][:guest_uuid]
which contains the UUID of the libvirt instance that the node is running
in. The purpose is to have a common place for this attribute independent
of the node's architecture (currently x86 ist storing this somewhere in
the "dmi" subtree and s390x in "s390x").