ohai: common ohai attribute for libvirt guest uuid
This adds a new attribute node[:crowbar_ohai][:libvirt][:guest_uuid]
which contains the UUID of the libvirt instance that the node is running
in. The purpose is to have a common place for this attribute independent
of the node's architecture (currently x86 ist storing this somewhere in
the "dmi" subtree and s390x in "s390x").

(cherry picked from commit 7548957de6b76c2760b76a72bd8bbfe2f48e738c)

This is only needed to fix the 6->7 upgrade, because the new attribute
is already used in the schema migration once we upgrade to Cloud7.