crowbar: Properly handle nodes rebooting at the end of chef
The reboot handler at the end of a chef run will cause /sbin/reboot,
which will result in the ssh session with chef-client to be suddenly
closed, resulting in what looks like a failure.

However, if we reach that reboot handler, then the chef run actually
completed since the handlers are run after the run itself.

This should fix the case of Xen compute nodes that cause the nova
proposal to fail to apply.