crowbar: Add delayed_job to enable easy background processing
This change adds delayed_job which enables us to create some tasks which
we can schedule in the background or we can execute only a single
method in the background. This reduces also the overhead we create with
the Thread.new approach. A nice side effect of this is that we can save
a lot of resources especially in big deployments because every new
thread needs the same amount of memory as the parent process.