fix breakage from conversion of upgrade state to symbols
PR #737 broke the test suite because it tried to call `#to_sym` on
`nil`, but Travis CI didn't notice because the unit test file was
incorrectly named.  (The file name will be fixed in the next commit.)

So instead of storing as strings and then symbolizing, we simply
store as symbols, eliminating the need for ugly calls to `#to_sym`
and `#deep_symbolize_keys`.