split loading off from #initialize
This will allow manually reloading state from the JSON file back into
the in-memory cache.  This means that a thread can ensure it always has
the latest state, rather than some outdated state cached in-memory,
since another thread could concurrently change the state.

This is crucially important for internal code which wants to change the
state, but it might also make sense to give external consumers the
choice of when they update their version of the state.