store upgrade state as YAML so we can use symbols
When representing the upgrade state, we want to use symbols instead of
strings for fixed enumerated values such as :pending or :running,
because that is the idiomatic Ruby way.  We also need to store strings
in the state file, for example errors messages or other details of the
progress of step, so we need a serialized file format which can
distinguish between strings and symbols.  Marshal can do this, but the
result is not human-readable.  JSON can't do this, which means you have
to complicate the code by carefully converting only specific keys and
values between symbols and strings in both directions.  This is
unavoidably brittle design.

(More information about the challenges of mixing and matching strings
with symbols is detailed here: http://stackoverflow.com/a/14872170/179332 )

Fortunately, YAML can serialize both symbols and strings, and it's even
more human-readable than JSON, so just use that instead.