add adminbackup api endpoint
it is necessary to have a single endpoint which stays the same over
consecutive releases. so in the next upgrade this can easily be re-
used.

if we would use /api/backups we have a conflict in cloud6 because
/api/backups is not backported, which means on cloud 6 the route
is located in /utils/backups.

there would also be a conflict with crowbarctl which exists because
crowbarctl is a shared rubygem among different cloud versions.
it would require a differentiation of cloud versions and endpoints
then.

to avoid all this, it is better to have a designated endpoint for
the admin backup in an upgrade case. It just adds minimal code
duplication, as the backups controller's create action does mostly
the same.

(cherry picked from commit 13a4080c3dc368b8216c10875d5fa70e7c2ef651)