Delete router namespaces to avoid router dups
Even if agents are migrated, we want to explicitly destroy the router
namespaces because in some upgrade cases, the routers remain active
when there are no valid destinations for the migration.

Example case, cluster with two nodes. The first one is upgraded, so
routers are moved to the second one. Once first node finishes the
upgrade, routers exist in the first and second nodes at once (because of
chef-client). Killing those routers in the pre-upgrade step of the
second node avoids that situation.