drop memoization and wrapper methods and cache check results once
since the upgrade model is now just a collection of class methods
we can't really use memoization anymore as the lifetime of the
class variable expands over multiple api requests that lead to
this code and this will always cause old cached results.

the wrapper methods are no longer needed, but the caching needs to
be done once per invocation of the checks method to enhance the
performance.