upgrade: let adminrepocheck be called multiple times
when the checks passed successfully, and the user runs them again
the status will advance from adminupgrade to database and so on.

it has to be handeled the same way as the prechecks, because the user
might call it multiple times.

so that when the checks passed successfully, and the user runs them
again the current step will not be changed.

(cherry picked from commit bf25c016b759fe226e729703a62e6652179f6a4d)