let adminrepocheck be called multiple times
when the checks passed successfully, and the user runs them again
the status will advance from adminupgrade to database and so on.

it has to be handeled the same way as the prechecks, because the user
might call it multiple times.

so that when the checks passed successfully, and the user runs them
again the current step will not be changed.