provisioner: Avoid crash with incomplete network mappings (bsc#990745)
We can hit a crash when a node has an incomplete admin network mapping.
This could happen if some ohai data is missing, or if the network
proposal is not providing a conduit map that is valid for a node.

If we keep crashing, then this actually breaks chef-client on the admin
server, which can then break DHCP, DNS and more for new nodes, or on
state transitions. However, there's nothing the admin server can do
about a node with wrong data, and we should avoid breaking the
deployment infrastructure because of this.

https://bugzilla.suse.com/show_bug.cgi?id=990745
(cherry picked from commit c00413c455a4d2f83b8557079f724f566e555912)