upgrade: force wicked config reload before upgrading (bsc#1011889)
If a node was never rebooted after the initial deployment and before
upgrading it, wicked seems to consider the interface configuration out
of sync with its current state (see e.g. "wicked ifcheck --change all"
output), which causes a wrong configuration to be applied when
openvswitch is upgraded.
We add a forced call to "wicked ifreload all" before starting the
actual upgrade to bring wicked in sync and workaround that issue.

The right fix is likely to add "wicked ifreload" call to the network
cookbooks when writing the ifcfg templates. That will follow with
another commit.

https://bugzilla.suse.com/show_bug.cgi?id=1011889