provisioner: Add autoyast self update option
AutoYaST can update itself before installing the OS. This patch enables
this feature and adds a field to the provisioner barclamp so that a
repository URL can be set by the user. The special token `<ADMINWEB>` resolves
to the IP:port of the admin node. An empty URL defaults to the SUSE
customer center (scc.suse.com).