update database configuration
in a fresh cloud setup this file gets overwritten by crowbar-init's
chef recipe, there is a template for that file.

nevertheless it should reflect the adapter that we have nowadays
which is postgres.

the other benefit of this patch is that when we use a development
environment the file gets overwritten by guard and the values are
still the correct defaults as we use from crowbar-init's template.