crowbar: Mark the proposal as not applied on saving
We mark it as applied when committing, but then we never reset it to not
applied when we do a change and just save. That results in the code
assuming that the saved proposal is always the latest one.

(cherry picked from commit cefacf692a5ed63ea178ee2601383fdfb6ed968e)