provisioner: Move uid/gid lower to avoid collision (bsc#998598)
For certain service users we fix their uid/gid combination to ensure
that the combination is consistent across multiple nodes. This is
necessary for interoperability with NFS for instance. By moving the
fixed uid/gid combinations to lower values we avoid (until there are
several hundred more service users in a default SLES installation)
collisions with the OS.