network: Run wicked ifreload only once per chef-client run (bsc#1012587)
This reworks the fix for https://bugzilla.suse.com/show_bug.cgi?id=1011889
to ensure that wicked ifreload is only execute at most once for a single
chef-client run. Previouly it was triggered once per updated ifcfg file
and that caused problems for certain configurations. E.g. when a single
chef-client run set up an "normal" physcial interface and a vlan
interface on top of that. (Happens e.g. in "dual" network setups when
allocating a public IP and when public and admin are on different
physical interfaces.)

https://bugzilla.suse.com/show_bug.cgi?id=1012587