upgrade: Move the upgrade step further before executing first chef-client
We want chef-client and crowbar_join services to be re-enabled once
the node has been upgraded. (They were disabled before upgrade.)

However, we still need to keep skipping pacemaker sync-marks,
because not all nodes of the cluster might be upgraded.
So we move the upgrade step to "done_os_upgrade", which will
indicate to crowbar-pacemaker that the sync-marks should be still skipped.