crowbar: Add infrastructure to put deployment config in data bag items
The goal of this is to allow barclamps to export the relevant config
bits to data bag items, so that nodes that are either not part of this
proposal but need config from that barclamp or nodes that require info
about other nodes in this proposal, can simply load the data bag item.

We also cache the config to avoid loading it multiple times per run.

This will help reduce the amount of search queries that are done from
the clients to the chef server.