upgrade: Ensure cinder controller runs shutdown of services as last node
As part of services shutdown, we're deleting cinder services from database.
For that, we need them to be stopped on every node before the code
that is executed from cinder controller does the deletion.

Also, we need to wait until this step finishes before the upgrade continues
with next steps.