cookbooks: Minimize search queries in recipes of admin server
We look for *:* multiple times, so we can cache that. It's worth
mentioning here that every time we run these queries, we do not need to
have the absolute latest attributes for a node, and a cache from the
beginning of the chef run is fine.

Also, the bind9 cookbook was searching for specific nodes to check if
they exist, while we already have the list of nodes.